fix: error when github server is down
When github is down, pr-police throws an error because it's expecting
the value data to be an array (calling .length on it).

This commit adds a check to see if data exists, if it doesn't returns
immediatly with a message stating that it could not fetch the issues, so
that users are warned correctly.