feat: isolates the configuration parsing logic
As the intention is to be able to support multiple projects from within
a single pr-cop instance (without breaking the compatibility with the
current configuration scheme), this commit moves the configuration
parsing/validation logic into a separate module.

The idea will then be to add a second source of configuration (i.e. a
file) and feed the already parsed configuration to the server function -
the extra logic will be implemented within this new configuration
module.