feat: load configuration from a file
This commit changes the `configuration` module in such a way that it's
now possible to provide a file path with the project configuration and
load the application configuration from there.

If the file path argument isn't provided, the previous "env var based"
behaviour is assumed (as to not break current applications using that
strategy).

If the file path is provided, the project related env vars are not
evaluated and instead the config file contents is used.